datagram api on darwin
* moved new code to interface/modules
* minimized build configs

going to add signature for ipv4 and epoll implementations
